Add Integration tests (#1)
* turning entity schema example into test

* changing requests to whylabs_client

* added README.md

* refactoring tests

* updating poetry

* fix mypy issue

* adding git actions environment

* adding env vars

* removing smoke test with unset variable

* ignoring jetbrains files

* addressing PR comments

+++ whylabs_toolkit/helpers/monitor_config.py
@@ -1,14 +1,9 @@
-import os
-import json
 import logging
 from typing import Any
 
-
-import requests
+from whylabs_client.exceptions import ApiValueError
 
 from whylabs_toolkit.helpers.utils import get_models_api
-from whylabs_toolkit.helpers.client import client
-from whylabs_client.exceptions import ApiValueError
 
 BASE_ENDPOINT = "https://api.whylabsapp.com"
 logger = logging.getLogger(__name__)
@@ -21,16 +16,9 @@ logger = logging.getLogger(__name__)
 
 
 def get_monitor_config(org_id: str, dataset_id: str) -> Any:
-    # TODO change to commented section once whylabs_client is updated
-    # api = get_models_api(org_id=org_id)
-    # monitor_config = api.get_monitor_config_v3(org_id=org_id, dataset_id=dataset_id)
-    get_monitor_config_url = f"v0/organizations/{org_id}/models/{dataset_id}/monitor-config/v3"
-    req_url = os.path.join(BASE_ENDPOINT, get_monitor_config_url)
-    resp = requests.get(
-        url=req_url,
-        headers={"accept": "application/json", "X-API-Key": client.configuration.api_key["ApiKeyAuth"]},
-    )
-    return json.loads(resp.content)
+    api = get_models_api()
+    monitor_config = api.get_monitor_config_v3(org_id=org_id, dataset_id=dataset_id)
+    return monitor_config
 
 
 def get_analyzer_ids(org_id: str, dataset_id: str, monitor_id: str) -> Any:
@@ -38,7 +26,7 @@ def get_analyzer_ids(org_id: str, dataset_id: str, monitor_id: str) -> Any:
     for item in monitor_config["monitors"]:
         if item["id"] == monitor_id:
             resp = item["analyzerIds"]
-    return resp
+            return resp
 
 
 def delete_monitor(org_id: str, dataset_id: str, monitor_id: str) -> None:
@@ -47,6 +35,8 @@ def delete_monitor(org_id: str, dataset_id: str, monitor_id: str) -> None:
         resp_monitor = api.delete_monitor(org_id=org_id, dataset_id=dataset_id, monitor_id=monitor_id)
         logger.debug(f"Deleted monitor with Resp:{resp_monitor}")
         analyzer_ids = get_analyzer_ids(org_id=org_id, dataset_id=dataset_id, monitor_id=monitor_id)
+        if analyzer_ids is None:
+            return
         for analyzer_id in analyzer_ids:
             resp_analyzer = api.delete_analyzer(org_id=org_id, dataset_id=dataset_id, analyzer_id=analyzer_id)
             logger.debug(f"Deleted analyzer with Resp:{resp_analyzer}")
